Move htmx.py into core
