+++ docs/static/documentation.txt
@@ -229,7 +229,7 @@ description: "Overview of the FastAPI webapp template architecture including dat
 This application uses a **hybrid Post-Redirect-Get (PRG) + HTMX** architecture. Every mutating endpoint supports both paths simultaneously:
 
 - **Non-HTMX path (PRG):** A standard browser form submission sends a POST request. On success the server issues a `303 See Other` redirect to a GET endpoint, which re-renders the full page with updated data. On error a full-page error template is returned.
-- **HTMX path:** When the browser sends the `HX-Request: true` header (added automatically by [htmx.org](https://htmx.org)), the same POST endpoint detects the header via `utils/htmx.py:is_htmx_request()` and instead returns a `200` HTML partial that HTMX swaps into the relevant section of the page. On error a toast partial is returned and swapped into `#toast-container` via out-of-band (OOB) swap.
+- **HTMX path:** When the browser sends the `HX-Request: true` header (added automatically by [htmx.org](https://htmx.org)), the same POST endpoint detects the header via `utils/core/htmx.py:is_htmx_request()` and instead returns a `200` HTML partial that HTMX swaps into the relevant section of the page. On error a toast partial is returned and swapped into `#toast-container` via out-of-band (OOB) swap.
 
 The HTMX rollout keeps the existing POST route contract intact — dedicated `PUT`/`PATCH`/`DELETE` routes may be introduced in a future iteration.
 
@@ -350,7 +350,7 @@ Toast partials are rendered from `templates/base/partials/toast.html` and inject
 
 ### HTMX request detection
 
-All HTMX-aware endpoints use the `is_htmx_request()` helper from `utils/htmx.py`:
+All HTMX-aware endpoints use the `is_htmx_request()` helper from `utils/core/htmx.py`:
 
 ```python
 def is_htmx_request(request: Request) -> bool:
@@ -843,7 +843,7 @@ Middleware functions are decorated with `@app.exception_handler(ExceptionType)` 
 Here's a middleware for handling the `PasswordValidationError` exception, which returns a toast partial for HTMX requests or a full error page for non-HTMX requests:
 
 ```python
-from utils.htmx import is_htmx_request
+from utils.core.htmx import is_htmx_request
 
 @app.exception_handler(PasswordValidationError)
 async def password_validation_exception_handler(request: Request, exc: PasswordValidationError):
